configs: Add gingerbred hardfork block to chain cofig (#188)
We need this hardfork to be able to determine how we should calculate
gas price for transactions and receipts returned over the RPC API.
Specifically we want to treat the pre gingerbread occurrences differently.

In L1 celo for pre gingerbread transactions when the associated state was
not available we would return null for the gas price, because we had no
way of knowing what the gas price was. Currently op-geth is returning the
gasFeeCap, which is misleading.