Add transfer precompile for token duality
Adds a new interface for celo precompiled contracts that accepts an
additional parameter (CeloPrecompiledContext) providing access to state
and EVM functionality.

Adds a struct that wraps exising precompiled contracts so that they
implement the new celo interface and ensures that existing contracts are
wrapped when accessed during the state transition.