Increase txpool inclusion (#332)
Given the one second block time of the celo chain, there is
significantly reduced need, compared to ethereum's twelve seconds, for
individual accounts to be able to include multiple transactions a single
block.

To account for this we can reduce the AccontSlots default. This makes
for a more inclusive txpool, capable of accepting transactions from a
more diverse range of accounts.

Even though it might make sense to divide the AccountSlots default by a
factor of 12, to match the account slots per second offered by ethereum,
doing so would reduce the Account slots below the max tx size of 4
slots. So we set the AccontSlots default at 4, to inrease inclusion
whilst still supporting large transactions.