Fix and update CI test workflows (#380)
* chore(ci): update foundry version to v1.1.0

* chore(ci): use new op-stack-base image runner

* chore: set default TERM in e2e test for tput

* chore(ci): determine npm version from package.json

* chore(ci): put setup in composite action

* chore: update e2e test js dependencies

* Adapt viem e2e test's asserts to upstream fix

Upstream viem recently merged a fix to the `estimateFeesPerGas`
function (https://github.com/wevm/viem/pull/3117),
and the recent updates to the e2e tests npm packages now includes the
fixed version of viem.

With this the asserted values in the viem e2e test are no longer a
reflection of what es expected to be returned from the viem library
and had to be adjusted.

* Loosen error check condition in e2e-test

Due to possible expansion of the responses error string in potential
proxy setups on a live RPC endpoint, the strict error check in the e2e
viem test caused the test to fail, although the error condition was
expected.