Add test case for USDC as a fee currency
This can be used to demonstrate the problem of getting a generic revert
message in an out-of-gas case because it uses a fee currency adapter
contract. See https://github.com/celo-org/op-geth/issues/346#issuecomment-3027651249

With the current values if works, but reducing the intrinsic gas to
40000 will exhibit the problem since that reduces that amount of gas
available to the creditGasFees function.